requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
backend==0.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
comm==0.2.2
crashtest==0.4.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
dulwich==0.21.7
ecdsa==0.19.0
executing==2.2.0
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.16.1
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
installer==0.7.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==24.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.1
more-itertools==10.5.0
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.2.2
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pip==25.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.19.1
pynput==1.7.7
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-json-logger==3.2.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.10.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
terminado==0.18.1
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
trove-classifiers==2024.10.21.16
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.28.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
xattr==1.1.0
